Moved country helper lookup to a new file for better encapsulation
