+++ app/utils/helper.py
@@ -48,14 +48,3 @@ def helper_get_latest_data_url(base_url: str) -> str:
         time_delta += 1
 
     return latest_base_url
-
-
-# Look up a country name from a country code
-def helper_lookup_country(country: str) -> str:
-    """ Look up a country name from a country code """
-    country_name = pycountry.countries.lookup(country).name # Select the first portion of str when , is found
-    if ',' in country_name:
-        country_name = country_name.split(',')[0]
-    elif ' ' in country_name:
-        country_name = country_name.split(' ')[-1]
-    return country_name
